requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
dj-database-url==2.0.0
Django==4.2.1
django-auto-logout==0.5.1
django-crispy-forms==2.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
openai==0.27.8
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
stripe==5.4.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.3
whitenoise==6.4.0
yarl==1.9.2
